requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
click==7.0
enum34==1.1.6             # via exif
exif==0.8.1
pillow==6.2.1
